requirements.txt
aisuite==0.1.6
altair==5.5.0
annotated-types==0.7.0
anthropic==0.30.1
anyio==4.7.0
attrs==24.3.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
filelock==3.16.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
groq==0.9.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.26.5
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.18.4
numpy==2.2.0
openai==1.57.4
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.29.1
pyarrow==18.1.0
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
streamlit==1.41.1
tenacity==9.0.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
